clipboard-apis/async-write-text-read-blobtext-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Async Clipboard writeText -> read (Blob/text) tests</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

function readBlobAsText(file) {
  return new Promise((resolve) => {
    const blobReader = new FileReader();
    blobReader.addEventListener("loadend", () => {
      resolve(blobReader.result);
    });
    blobReader.readAsText(file);
  });
};

promise_test(async t => {
  const input = "Clipboard writeText -> read (Blob/text) test データ";

  await navigator.clipboard.writeText(input);
  const output = await navigator.clipboard.read();

  const textOutput = await readBlobAsText(output);

  assert_equals(output.type, "text/plain");
  assert_equals(textOutput, input);
}, "Verify write and read clipboard (DOMString)");
</script>
Note: This is a manual test because it writes/reads to the shared system
clipboard and thus cannot be run async with other tests that might interact
with the clipboard.
